+++ IPython/html/nbconvert/tests/test_nbconvert_handlers.py
@@ -1,3 +1,4 @@
+# coding: utf-8
 import io
 import json
 import os
@@ -95,10 +96,3 @@ class APITest(NotebookTestBase):
         r = self.nbconvert_api.from_post(format='python', nbmodel=nbmodel)
         self.assertIn(u'text/x-python', r.headers['Content-Type'])
         self.assertIn(u'print(2*6)', r.text)
-
-    def test_list_formats(self):
-        formats = self.nbconvert_api.list_formats().json()
-        self.assertIsInstance(formats, dict)
-        self.assertIn('python', formats)
-        self.assertIn('html', formats)
-        self.assertEqual(formats['python']['output_mimetype'], 'text/x-python')
